requirements.txt
discord.py
transformers
torch
pillow
requests
numpy
nudity-detector

